==> Step0_test.py <==
import Step0_test
import Step1_BalloonOriginalData
import Step2_BalloonDataCleanWithWind
import Step3_FetchPlaneData
import Step4_BalloonForecast
import Step5_dangerousballoons
import Step6_visualizemap   

def my_function():
    print("test")
